ENH: Basic linear intensity correction, changes in volumetric recon pipeline
- Minor changes in volumetric recon pipeline
- Added linear intensity correction to data preprocessing
- Segmentation propagation with more


Former-commit-id: 8cc2e5045c22a4dfd86d54fde2ee21c1dfc4ed48

+++ src/py/utilities/StackMaskMorphologicalOperations.py
@@ -101,7 +101,7 @@ class StackMaskMorphologicalOperations(object):
 
             shape = self._mask_sitk.GetSize()
             N_slices = shape[2]
-            nda_mask = np.zeros(shape[::-1])
+            nda_mask = np.zeros(shape[::-1], dtype=np.uint16)
 
             for i in range(0, N_slices):
                 slice_mask_sitk = self._mask_sitk[:,:,i:i+1]
